# protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Family Robot Local API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/speech/preview": {
      "post": {
        "summary": "Preview",
        "operationId": "preview_v1_speech_preview_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Speak"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/speech/recognize": {
      "post": {
        "summary": "Recognize",
        "operationId": "recognize_v1_speech_recognize_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_v1_speech_recognize_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/books/recognize": {
      "post": {
        "summary": "Recognize Book",
        "operationId": "recognize_book_v1_books_recognize_post",
        "parameters": [
          {
            "name": "imageAgeMs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Imageagems"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_book_v1_books_recognize_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/turns": {
      "post": {
        "summary": "Turn",
        "operationId": "turn_v1_turns_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Turn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_v1_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pairing": {
      "post": {
        "summary": "Invite",
        "operationId": "invite_v1_pairing_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pairing/claim": {
      "post": {
        "summary": "Claim",
        "operationId": "claim_v1_pairing_claim_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairClaim"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/resources": {
      "post": {
        "summary": "Create",
        "operationId": "create_v1_resources_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listing",
        "operationId": "listing_v1_resources_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Kind"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Language"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "State"
            }
          },
          {
            "name": "favorite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Favorite"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_recognize_book_v1_books_recognize_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_recognize_book_v1_books_recognize_post"
      },
      "Body_recognize_v1_speech_recognize_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_recognize_v1_speech_recognize_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 30,
            "title": "Label",
            "default": ""
          },
          "chapter": {
            "type": "string",
            "maxLength": 100,
            "title": "Chapter",
            "default": ""
          },
          "sourceAsset": {
            "type": "string",
            "maxLength": 64,
            "title": "Sourceasset",
            "default": ""
          },
          "sourcePage": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 0.0,
            "title": "Sourcepage",
            "default": 0
          },
          "sourceRotation": {
            "type": "integer",
            "enum": [
              0,
              90,
              180,
              270
            ],
            "title": "Sourcerotation",
            "default": 0
          },
          "extractionError": {
            "type": "string",
            "maxLength": 80,
            "title": "Extractionerror",
            "default": ""
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "qualityWarnings": {
            "items": {
              "type": "string",
              "enum": [
                "BLANK",
                "DUPLICATE",
                "LOW_CONFIDENCE",
                "OCR_FAILED"
              ]
            },
            "type": "array",
            "maxItems": 4,
            "title": "Qualitywarnings"
          },
          "text": {
            "type": "string",
            "maxLength": 20000,
            "title": "Text",
            "default": ""
          },
          "reviewed": {
            "type": "boolean",
            "title": "Reviewed",
            "default": false
          },
          "skip": {
            "type": "boolean",
            "title": "Skip",
            "default": false
          },
          "pronunciation": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "maxProperties": 100,
            "title": "Pronunciation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Page"
      },
...
      "PairClaim": {
        "properties": {
          "invite": {
            "type": "string",
            "maxLength": 200,
            "minLength": 20,
            "title": "Invite"
          },
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "invite",
          "name"
        ],
        "title": "PairClaim"
      },
...
      "Register": {
        "properties": {
          "invite": {
            "type": "string",
            "maxLength": 200,
            "minLength": 20,
            "title": "Invite"
          },
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "invite",
          "name"
        ],
        "title": "Register"
      },
      "ResourceCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "book",
              "story",
              "song",
              "dialogue"
            ],
            "title": "Kind"
          },
          "draft": {
            "$ref": "#/components/schemas/ResourceDraft"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "draft"
        ],
        "title": "ResourceCreate"
      },
      "ResourceDraft": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Aliases"
          },
          "language": {
            "type": "string",
            "enum": [
              "zh",
              "en",
              "bilingual"
            ],
            "title": "Language",
            "default": "zh"
          },
          "edition": {
            "type": "string",
            "maxLength": 100,
            "title": "Edition",
            "default": ""
          },
          "author": {
            "type": "string",
            "maxLength": 100,
            "title": "Author",
            "default": ""
          },
          "publisher": {
            "type": "string",
            "maxLength": 100,
            "title": "Publisher",
            "default": ""
          },
          "isbn": {
            "type": "string",
            "maxLength": 20,
            "title": "Isbn",
            "default": ""
          },
          "source": {
            "type": "string",
            "maxLength": 500,
            "title": "Source",
            "default": "家长提供"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Tags"
          },
          "minAge": {
            "type": "integer",
            "maximum": 18.0,
            "minimum": 0.0,
            "title": "Minage",
            "default": 3
          },
          "maxAge": {
            "type": "integer",
            "maximum": 18.0,
            "minimum": 0.0,
            "title": "Maxage",
            "default": 12
          },
          "englishLevel": {
            "type": "string",
            "enum": [
              "beginner",
              "basic",
              "intermediate"
            ],
            "title": "Englishlevel",
            "default": "beginner"
          },
          "favorite": {
            "type": "boolean",
            "title": "Favorite",
            "default": false
          },
          "coverAsset": {
            "type": "string",
            "maxLength": 64,
            "title": "Coverasset",
            "default": ""
          },
          "audioAsset": {
            "type": "string",
            "maxLength": 64,
            "title": "Audioasset",
            "default": ""
          },
          "complete": {
            "type": "boolean",
            "title": "Complete",
            "default": false
          },
          "excerpt": {
            "type": "string",
            "maxLength": 200,
            "title": "Excerpt",
            "default": ""
          },
          "auditioned": {
            "type": "boolean",
            "title": "Auditioned",
            "default": false
          },
          "voice": {
            "$ref": "#/components/schemas/Voice"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/Page"
            },
            "type": "array",
            "maxItems": 300,
            "title": "Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ResourceDraft"
      },
...
      "Speak": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 600,
            "minLength": 1,
            "title": "Text"
          },
          "voice": {
            "$ref": "#/components/schemas/Voice"
          },
          "story": {
            "type": "boolean",
            "title": "Story",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "Speak"
      },
      "Turn": {
        "properties": {
          "sessionId": {
            "type": "string",
            "maxLength": 80,
            "minLength": 16,
            "title": "Sessionid"
          },
          "text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Text"
          },
          "image": {
            "type": "string",
            "maxLength": 1500000,
            "title": "Image",
            "default": ""
          },
          "imageAgeMs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Imageagems",
            "default": 0
          },
          "visualRequest": {
            "type": "boolean",
            "title": "Visualrequest",
            "default": false
          },
          "currentResourceId": {
            "type": "string",
            "maxLength": 80,
            "title": "Currentresourceid",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sessionId",
          "text"
        ],
        "title": "Turn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Voice": {
        "properties": {
          "zh": {
            "type": "string",
            "maxLength": 80,
            "title": "Zh",
            "default": "default"
          },
          "en": {
            "type": "string",
            "maxLength": 80,
            "title": "En",
            "default": "default"
          },
          "story": {
            "type": "string",
            "maxLength": 80,
            "title": "Story",
            "default": "default"
          },
          "speed": {
            "type": "number",
            "maximum": 1.3,
            "minimum": 0.7,
            "title": "Speed",
            "default": 1.0
          },
          "volume": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Volume",
            "default": 0.5
          },
          "style": {
            "type": "string",
            "enum": [
              "neutral",
              "gentle",
              "cheerful",
              "storytelling",
              "soothing"
            ],
            "title": "Style",
            "default": "gentle"
          },
          "storyStyle": {
            "type": "string",
            "enum": [
              "neutral",
              "gentle",
              "cheerful",
              "storytelling",
              "soothing"
            ],
            "title": "Storystyle",
            "default": "storytelling"
          },
          "instruction": {
            "type": "string",
            "maxLength": 200,
            "title": "Instruction",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Voice"
      }
    }
  }
}